Fix compatibility issues to pass test suite

+++ include/chip8/cpu.h
@@ -12,7 +12,7 @@ typedef struct
     uint16_t I;
     uint16_t pc;
     uint16_t sp;
-    uint8_t stack[16];
+    uint16_t stack[16];
     uint8_t delay_timer;
     uint8_t sound_timer;
 } chip8_cpu;
